+++ src/main/java/com/kambr/parser/onex/spo/dataClasses/FromAndPrice.kt
@@ -1,3 +1,5 @@
 package com.kambr.parser.onex.spo.dataClasses
 
-data class FromAndPrice(val from: Int, val price: Double)
+import java.math.BigDecimal
+
+data class FromAndPrice(val from: Int, val price: BigDecimal)
